services/asr/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic-settings==2.10.1
redis==6.4.0
psycopg[binary]==3.2.9
psycopg-pool==3.2.6
boto3==1.40.18
prometheus-client==0.22.1
python-multipart==0.0.20
numpy==2.3.2
pydub==0.25.1
faster-whisper==1.2.0
